+++ src/kaxanuk/data_curator/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.44.0'
+__version__ = '0.45.0'
 __parameters_format_version__ = '0.40.0'
 
 __package_name__ = 'kaxanuk.data_curator'